Changed to use IBM's model
Fixes #1

+++ src/index.ts
@@ -1,8 +1,9 @@
-import { pythonBridge, PythonBridge } from 'python-bridge';
 import * as fs from 'fs';
 import * as Telegraf from 'telegraf';
 import { MainService } from './main-service';
 import { Actions } from './actions';
+import { fetch, Headers } from 'cross-fetch';
+import { User } from 'telegraf/typings/telegram-types';
 
 const schinquirer = require("@luismayo/schinquirer");
 
@@ -16,50 +17,69 @@ async function main() {
         fs.writeFile('./conf/conf.json', JSON.stringify(MainService.conf), {encoding: 'utf-8'}, () => {});
     }
     const bot = new Telegraf.default(MainService.conf.token);
-    const python = pythonBridge({
-        python: MainService.conf.pythonInterpreter,
-        stdio: ['ignore', 'pipe', 'ignore']
-    });
-    python.ex`from hatesonar import Sonar`;
-    python.ex`import json`;
-    python.ex`sonar = Sonar()`;
-
     bot.start(ctx => ctx.reply('I will moderate delete hate messages and/or kick users. Please make me an admin'));
     bot.use(ctx => {
-        const text = ctx.message.text || ctx.message.caption;
-        if (text && ctx.message.from) {
-            checkText(text, python, ctx).then((result) => {
-                if (result === Actions.KICK || result === Actions.DELETE) {
-                    ctx.deleteMessage();
-                }
-                if (result === Actions.KICK) {
-                    ctx.kickChatMember(ctx.from.id);
-                }
-            });
+        if (ctx.message) {
+            const text = ctx.message.text || ctx.message.caption;
+            if (text && ctx.message.from) {
+                checkText(text, ctx).then((result) => {
+                    if (result === Actions.KICK || result === Actions.DELETE) {
+                        ctx.deleteMessage();
+                        if (result === Actions.DELETE) {
+                            ctx.reply(`⚠ ${makeUserLink(ctx.from)} watch your language or you'll be banned`, {parse_mode: 'Markdown'});
+                        }
+                    }
+                    if (result === Actions.KICK) {
+                        ctx.kickChatMember(ctx.from.id);
+                    }
+                });
+            }
         }
     });
     bot.launch();
 }
 
-async function checkText(text: string, python: PythonBridge, ctx: Telegraf.Context) {
+async function checkText(text: string, ctx: Telegraf.Context) {
     // const result = await python.ex`json.dumps(sonar.ping(text=${text}))`;
-    const result = await python`sonar.ping(text=${text})`;
+    const headers = new Headers();
+    headers.set('Content-Type', 'application/json');
+    const result = await (await fetch(MainService.conf.serverUrl + 'model/predict', {method: 'POST', headers: headers, body: JSON.stringify({text: [text]})})).json();
     if (MainService.conf.debugMode) {
-        ctx.reply(JSON.stringify(result));
+        ctx.reply(JSON.stringify(result, null, 2));
     }
-    if (result.top_class !== 'neither') {
-        const hate = result.classes.find(item => item.class_name === 'hate_speech');
-        const offensive = result.classes.find(item => item.class_name === 'offensive_language');
-        if (hate.confidence >= MainService.conf.kickConfidenceHate || offensive.confidence >= MainService.conf.kickConfidenceOffensive) {
+
+    const prediction = result.results[0].predictions;
+
+    /*
+    "toxic": 0.00043087179074063897,
+        "severe_toxic": 0.00012918129505123943,
+        "obscene": 0.00015268969582393765,
+        "threat": 0.00011289756366750225,
+        "insult": 0.00013602284889202565,
+        "identity_hate": 0.00014075539365876466 */
+        if (prediction.toxic >= MainService.conf.toxicKickConfidence
+            || prediction.severe_toxic >= MainService.conf.severeToxicKickConfidence
+            || prediction.obscene >= MainService.conf.obsceneKickConfidence
+            || prediction.threat >= MainService.conf.threatKickConfidence
+            || prediction.insult >= MainService.conf.insultKickConfidence
+            || prediction.identity_hate >= MainService.conf.hateKickConfidence) {
             console.log(new Date().toISOString() + ' kick');
             return Actions.KICK;
         }
-        if (hate.confidence >= MainService.conf.deleteConfidenceHate || offensive.confidence >= MainService.conf.deleteConfidenceOffensive) {
-            console.log(new Date().toISOString() + ' delete');
+        if (prediction.toxic >= MainService.conf.toxicDeleteConfidence
+            || prediction.severe_toxic >= MainService.conf.severeToxicDeleteConfidence
+            || prediction.obscene >= MainService.conf.obsceneDeleteConfidence
+            || prediction.threat >= MainService.conf.threatDeleteConfidence
+            || prediction.insult >= MainService.conf.insultDeleteConfidence
+            || prediction.identity_hate >= MainService.conf.hateDeleteConfidence) {
+            console.log(new Date().toISOString() + ' kick');
             return Actions.DELETE;
         }
-    }
     return Actions.NOTHING;
+}
+
+function makeUserLink(usr: User) {
+    return `[${usr.first_name}](tg://user?id=${usr.id})`
 }
 
 main().then(() => {});
